accessible/tests/mochitest/events/docload/test_docload_shutdown.html

<html>

<head>
  <title>Accessible events testing for document</title>

  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="../../common.js"></script>
  <script type="application/javascript"
          src="../../role.js"></script>
  <script type="application/javascript"
          src="../../events.js"></script>

  <script type="application/javascript">
    async function doTests() {
      if (Services.appShell.hasHiddenWindow) {
        // Front end stuff sometimes likes to stuff things in the hidden window(s)
        // in which case we should repress all accessibles for those.

        // Try to create an accessible for the hidden window's document.
        let doc = Services.appShell.hiddenDOMWindow.document;
        let hiddenDocAcc = gAccService.getAccessibleFor(doc);
        ok(!hiddenDocAcc, "hiddenDOMWindow should not have an accessible.");
      }

      // Get application root accessible.
      let docAcc = getAccessible(document);
      let rootAcc = null;
      while (docAcc) {
        rootAcc = docAcc;
        try {
          docAcc = docAcc.parent;
        } catch (e) {
          ok(false, `Can't get parent for ${prettyName(docAcc)}`);
          throw e;
        }
      }

      const aURL = "../../events/docload/docload_wnd.html";
      const reordered = waitForEvent(EVENT_REORDER, rootAcc);
      const hidden = waitForEvent(EVENT_HIDE, "iframe");
      let dialog = window.browsingContext.topChromeWindow.openDialog(aURL);
      await reordered;
      let iframe = (await hidden).DOMNode;

      testAccessibleTree(rootAcc, {
        role: ROLE_APP_ROOT,
        children: [
          { role: ROLE_CHROME_WINDOW },
          { role: ROLE_CHROME_WINDOW },
        ],
      });
      // After timeout after event hide for iframe was handled the document
      // accessible for iframe's document should no longer be in cache.
      ok(!isAccessibleInCache(iframe.contentDocument),
        "The document accessible for iframe is in cache still after iframe hide!");
      ok(isAccessibleInCache(dialog.document),
        `The document accessible for '${aURL}' is not in cache!`);

      const focused = waitForEvent(EVENT_FOCUS, getAccessible(document));
      dialog.close();
      await focused;

      ok(!isAccessibleInCache(dialog.document),
        "The document accessible for dialog is in cache still!");
      // After the window is closed all alive subdocument accessibles should
      // be shut down.
      ok(!isAccessibleInCache(iframe.contentDocument),
        "The document accessible for iframe is in cache still!");

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTests);
  </script>
</head>

<body>

  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=571459"
     title="Shutdown document accessible when presshell goes away">
    Mozilla Bug 571459
   </a>

  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>
</body>
</html>
